Create addon channel in non-storyStoreV7
