+++ packages/builder-vite/codegen-modern-iframe-script.ts
@@ -1,6 +1,6 @@
 import { loadPreviewOrConfigFile } from '@storybook/core-common';
 import { normalizePath } from 'vite';
-import { virtualStoriesFile } from './virtual-file-names';
+import { virtualStoriesFile, virtualAddonSetupFile } from './virtual-file-names';
 import type { ExtendedOptions } from './types';
 
 export async function generateModernIframeScriptCode(options: ExtendedOptions) {
@@ -21,37 +21,20 @@ export async function generateModernIframeScriptCode(options: ExtendedOptions) {
    */
   // language=JavaScript
   const code = `
-    import global from 'global';
-
     import { composeConfigs, PreviewWeb } from '@storybook/preview-web';
     import { ClientApi } from '@storybook/client-api';
-    import { addons } from '@storybook/addons';
-    import createPostMessageChannel from '@storybook/channel-postmessage';
-    import createWebSocketChannel from '@storybook/channel-websocket';
-
-    import { importFn } from '${storiesFilename}';
-
-    const { SERVER_CHANNEL_URL } = global;
+    import '${virtualAddonSetupFile}';
+    import { importFn } from '${virtualStoriesFile}';
 
     const getProjectAnnotations = async () =>
       composeConfigs(await Promise.all([${configEntries
         .map((configEntry) => `import('${configEntry}')`)
         .join(',\n')}]));
 
-    const channel = createPostMessageChannel({ page: 'preview' });
-    addons.setChannel(channel);
-
-    if (SERVER_CHANNEL_URL) {
-      const serverChannel = createWebSocketChannel({ url: SERVER_CHANNEL_URL });
-      addons.setServerChannel(serverChannel);
-      window.__STORYBOOK_SERVER_CHANNEL__ = serverChannel;
-    }
-
     const preview = new PreviewWeb();
 
     window.__STORYBOOK_PREVIEW__ = preview;
     window.__STORYBOOK_STORY_STORE__ = preview.storyStore;
-    window.__STORYBOOK_ADDONS_CHANNEL__ = channel;
     window.__STORYBOOK_CLIENT_API__ = new ClientApi({ storyStore: preview.storyStore });
 
     preview.initialize({ importFn, getProjectAnnotations });
